Adding Ext.model.Resource and Ext.store.Resources
Adding a Resource model that represents the Agency/Service combo that the Teen Link staff was taking about.

This is a temporary model and store that we'll be using to get started with this application.

Additionally we have to make sure the application is aware of the new model and store.

Signed-off-by: Ross Hattori <git@exstasis.net>