Change the case of the topic title testing data
It makes more sense to have the titles all lower case so we can do real-time text transformations on the string for display. It also helps show what is an acronym but doesn't solve the presentation problem of how we property transform text on the fly so it displays correctly.