Start of the sqlalchemy (SA) ORM modeling
This commit represents the first steps towards the new import process
that I want to create. The previous python scripts are error-prone and
I could see problems when starting to write data validations into the
code.

Instead I've decided to make a python module that clearly defines the
expected fields and types of the data and allows clear places for the
munging logic. By defining the model first--something we couldn't do
when we initially started this project--we can separate the code used to
import data from different sources, the validation/cleanup code, the
export code, configuration, and cli tool for easier maintainability.

Additionally using SA allows me to easily make a flask-based API around
the database if we need it later.